Fix for Issue 214; work-around for probable bug in Honeycomb.
It seems that Honeycomb releases cursors in Activitiy objects which
have been passed to startManagingObjects() but the underlying users of
those cursors (eg. intrinsic Android adapters) do not expect them to
be deleted.

This patch replaces startManagingCursor() functionality.

Unfortunately Google have not released Android 3.0 source yet, so it
is hard to tell the cause.