Fix bug when importing or adding duplicate anthology titles
Anthology titles that differ only in a leading space may exist in an
export file. These are trimmed on import and cause a duplicate index
error. New code checks for existing title and does not add if already
present.

Further, in prior versions, adding a second anthology title that exactly
matched an existng one would crash the app