Reduce risk of losing book list expanded nodes on orientation change
The emulator (and maybe other devices) fire two configuration changes
in rapid succession when rotating from landscape to portrait. This
previously mean that the onPause() would be fired before the list had
build, and the wrong expansion settings would be saved.

Now, expanded nodes are saved when they are changed, and position data
is only saved if the list has been built once.