Deal more gracefull with missing or currupt covers db
The covers DB is not *essential*; it is just an optimization. Testing
reveled in can become corrupt and when it does, we just need to silently
recreate it.

If we still can't open it, then we mark is as unavailable and don't
attempt to retry until the next app instance is started.