Full backup is (almost) here
This version does full backup/restore, but has limitations which are
actively being worked on:

- you can not choose where to save the backup, or change the file name
(currently it goes into the 'bookCatalogue' directory, with the name
'BookCatalogue.bcbk'. Fixing this is MANDATORY for release.

- it does not really do a proper OVERWRITE restore; when you restore the
backup it gets added/merged with existing data (the same as importing an
export file in another directory).  Fixing this is IMPORTANT for
release.

- You can not choose what to backup (eg. "no covers"); this will
probably NOT be in the first release