Bug fixes for evil design flaws and some renaming
Original version of FileBrowser was too slack in the way it handled
listeners. This version strictly adheres to the "fragments communicate
via the Activity" paradigm, resulting in fewer crashes on rotation etc.

This version also more cleanly encapsulates view handling so that we
don't need a new fragment for each file-type browser, but we do need a
FileChooser Activity, a FileDetails object and a FileLister.