Bug fix for Android 2.1 and double-dereferencing SQLiteDatabase objects
There seems to be a subtle bug in Android 2.1 that results in the
reference count of SQLiteDatabase objects being decremented too many
times.

The simplest way to see this is to:

- comment out the onPause() code in BookEdit
- start the app
- go to "My Books"
- click on a book
- press 'back' until you get back to MainMenu
- select "Ny Books" 

This will crash in Android 2.1 with an "already closed" message on the
database (an SQLiteClosable).

There may well be other similar sources of this problem.

A more drastic solution would be to call acquireReference() on the
database 100 times, then use reflection to make sure the number never
drops below 100...but it's probably best to avoid such techniques.

With luck, we can just discontinue support for Android 2.1 Soon(tm).