Check if GetList queue is terminating, and discard resources if so
This is *not* a complete fix for the Android 2.1 statement problem, but
it reduces the impact.

There is still a chance that a queue may have finish() called after a
task has completed but before onFinish() is run. Which is why we need
some kind of onTerminate() handler, or a flag to onFinish().