Fix duplicate series problem
Series of the form "SomeName(2)" and "SomeName(2.0)" are different, but
becuase series_num in the database was an integer, the 2.0 was being
coerced into a "2", resulting in a duplicate index error.

The series_num column has been changed to 'text'. This should not affect
booklist sorting as the existing code casts it to a float for sorting
purposes.