As released with changes to fix startup process and ensure move-file message displays, added uses-feature (camera) to manifest
In retrospect this was not the ideal way to perform special upgrad steps. A better approach would have been to create a whole new activity to function as a dialog. That activity would manage the upgrade and THEN start the main activity. This has the advantage of an arms-length approach from the main app functions and thus would rely less on 'standard' app features (and interfere with them less).